# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/organizations": {
      "post": {
        "description": "Le créateur devient automatiquement OWNER. Un `slug` unique est dérivé du nom commercial (ou de la raison sociale) et sert ensuite à la numérotation des reçus : il n'est plus modifiable.",
        "operationId": "OrganizationsController_create",
        "parameters": [
          {
            "name": "X-Organization-Id",
            "in": "header",
            "required": false,
            "description": "Organisation courante. Obligatoire sur les routes d'organisation ; ignoré ailleurs.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrganizationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationDto"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Créer une organisation",
        "tags": ["Organisations"]
      }
    },
    "/v1/organizations/{id}": {
      "get": {
        "operationId": "OrganizationsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Organization-Id",
            "in": "header",
            "description": "Organisation courante. Positionne `app.current_organization_id` (RLS).",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationDto"
                }
              }
            }
          },
          "404": {
            "description": "ORG.NOT_MEMBER",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Détail d'une organisation",
        "tags": ["Organisations"]
      },
      "patch": {
        "operationId": "OrganizationsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Organization-Id",
            "in": "header",
            "description": "Organisation courante. Positionne `app.current_organization_id` (RLS).",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrganizationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mettre à jour une organisation",
        "tags": ["Organisations"]
      }
    },
    "/v1/organizations/{id}/settings": {
      "get": {
        "operationId": "OrganizationsController_getSettings",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Organization-Id",
            "in": "header",
            "description": "Organisation courante. Positionne `app.current_organization_id` (RLS).",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationSettingsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Paramètres de l'organisation",
        "tags": ["Organisations"]
      },
      "patch": {
        "description": "La devise est verrouillée à XAF par contrainte SQL et ne peut pas être modifiée.",
        "operationId": "OrganizationsController_updateSettings",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Organization-Id",
            "in": "header",
            "description": "Organisation courante. Positionne `app.current_organization_id` (RLS).",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrganizationSettingsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationSettingsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mettre à jour les paramètres",
        "tags": ["Organisations"]
      }
    },
    "/v1/organizations/{id}/members": {
      "get": {
        "operationId": "OrganizationsController_listMembers",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Organization-Id",
            "in": "header",
            "description": "Organisation courante. Positionne `app.current_organization_id` (RLS).",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberListDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lister les membres",
        "tags": ["Organisations"]
      }
    },
    "/v1/organizations/{id}/members/{memberId}": {
      "patch": {
        "description": "Le dernier OWNER actif ne peut pas être rétrogradé : 409 `ORG.LAST_OWNER`.",
        "operationId": "OrganizationsController_changeMemberRole",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "memberId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Organization-Id",
            "in": "header",
            "description": "Organisation courante. Positionne `app.current_organization_id` (RLS).",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMemberRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberDto"
                }
              }
            }
          },
          "409": {
            "description": "ORG.LAST_OWNER",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Changer le rôle d'un membre",
        "tags": ["Organisations"]
      },
      "delete": {
        "description": "Le dernier OWNER actif ne peut pas être retiré : 409 `ORG.LAST_OWNER`.",
        "operationId": "OrganizationsController_removeMember",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "memberId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Organization-Id",
            "in": "header",
            "description": "Organisation courante. Positionne `app.current_organization_id` (RLS).",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "409": {
            "description": "ORG.LAST_OWNER",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Retirer un membre",
        "tags": ["Organisations"]
      }
    },
    "/v1/organizations/{id}/invitations": {
      "get": {
        "operationId": "OrganizationsController_listInvitations",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Organization-Id",
            "in": "header",
            "description": "Organisation courante. Positionne `app.current_organization_id` (RLS).",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationListDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lister les invitations",
        "tags": ["Organisations"]
      },
      "post": {
        "description": "Un jeton de 32 octets est envoyé par SMS ; seul son condensat SHA-256 est stocké. L'invitation est nominative : seul le titulaire du numéro peut l'accepter.",
        "operationId": "OrganizationsController_createInvitation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Organization-Id",
            "in": "header",
            "description": "Organisation courante. Positionne `app.current_organization_id` (RLS).",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInvitationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Inviter un collaborateur",
        "tags": ["Organisations"]
      }
    },
    "/v1/organizations/{id}/invitations/{invitationId}": {
      "delete": {
        "operationId": "OrganizationsController_revokeInvitation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "invitationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Organization-Id",
            "in": "header",
            "description": "Organisation courante. Positionne `app.current_organization_id` (RLS).",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Annuler une invitation",
        "tags": ["Organisations"]
      }
    },
...
  },
  "info": {
    "title": "Immodesk API",
    "description": "API de la plateforme Immodesk (gestion immobilière, Congo-Brazzaville).\n\n**Conventions**\n- Préfixe : `/v1`. Dates ISO 8601 UTC, montants entiers XAF (BigInt sérialisé en chaîne).\n- Authentification : `Authorization: Bearer <accessToken>` (JWT, 15 minutes).\n- Contexte d'organisation : en-tête `X-Organization-Id` obligatoire sur toute route d'organisation.\n- Erreurs : `{ code, message, details }`, codes stables `DOMAINE.RAISON`, messages en français (fr-CG).\n- Toute ressource appartenant à une autre organisation répond **404**, jamais 403.\n- Pagination par curseur : `?limit=50&cursor=...` → `{ items, pageInfo }`.\n- Idempotence : en-tête facultatif `Idempotency-Key` sur les POST.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000/v1",
      "description": "Développement local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Jeton d'accès (15 min)."
      }
    },
    "schemas": {
...
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "IAM.OTP_INVALID",
            "description": "Code stable `DOMAINE.RAISON`."
          },
          "message": {
            "type": "string",
            "example": "Code incorrect.",
            "description": "Message en français (fr-CG)."
          },
          "details": {
            "type": "object",
            "example": {
              "remainingAttempts": 3
            }
          }
        },
        "required": ["code", "message", "details"]
      },
...
      "OrganizationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "type": "string",
            "enum": ["AGENCY", "INDEPENDENT_LANDLORD", "INDEPENDENT_MANAGER"]
          },
          "legalName": {
            "type": "string"
          },
          "tradeName": {
            "type": "string",
            "nullable": true
          },
          "slug": {
            "type": "string"
          },
          "city": {
            "type": "string",
            "example": "Brazzaville"
          },
          "district": {
            "type": "string",
            "nullable": true
          },
          "contactPhone": {
            "type": "string"
          },
          "contactEmail": {
            "type": "string",
            "nullable": true
          },
          "logoUrl": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": ["ACTIVE", "SUSPENDED"]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "type",
          "legalName",
          "tradeName",
          "slug",
          "city",
          "district",
          "contactPhone",
          "contactEmail",
          "logoUrl",
          "status",
          "createdAt"
        ]
      },
...
      "CreateOrganizationDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": ["AGENCY", "INDEPENDENT_LANDLORD", "INDEPENDENT_MANAGER"]
          },
          "legalName": {
            "type": "string",
            "example": "Agence Mpila Immo SARL"
          },
          "tradeName": {
            "type": "string",
            "example": "Agence Mpila Immo"
          },
          "city": {
            "type": "string",
            "example": "Brazzaville"
          },
          "district": {
            "type": "string",
            "example": "Mpila"
          },
          "contactPhone": {
            "type": "string",
            "example": "+242066000001"
          },
          "contactEmail": {
            "type": "string",
            "example": "contact@mpila-immo.cg"
          }
        },
        "required": ["type", "legalName", "city", "contactPhone"]
      },
      "UpdateOrganizationDto": {
        "type": "object",
        "properties": {
          "legalName": {
            "type": "string"
          },
          "tradeName": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "district": {
            "type": "string"
          },
          "contactPhone": {
            "type": "string"
          },
          "contactEmail": {
            "type": "string"
          },
          "logoDocumentId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "OrganizationSettingsDto": {
        "type": "object",
        "properties": {
          "defaultPaymentDueDay": {
            "type": "number",
            "example": 5,
            "minimum": 1,
            "maximum": 28
          },
          "timezone": {
            "type": "string",
            "example": "Africa/Brazzaville"
          },
          "currency": {
            "type": "string",
            "enum": ["XAF"],
            "example": "XAF"
          },
          "defaultGraceDays": {
            "type": "number",
            "example": 5
          },
          "receiptFooterText": {
            "type": "string",
            "nullable": true
          },
          "whatsappEnabled": {
            "type": "boolean"
          },
          "smsEnabled": {
            "type": "boolean"
          }
        },
        "required": [
          "defaultPaymentDueDay",
          "timezone",
          "currency",
          "defaultGraceDays",
          "receiptFooterText",
          "whatsappEnabled",
          "smsEnabled"
        ]
      },
      "UpdateOrganizationSettingsDto": {
        "type": "object",
        "properties": {
          "defaultPaymentDueDay": {
            "type": "number",
            "minimum": 1,
            "maximum": 28
          },
          "timezone": {
            "type": "string",
            "example": "Africa/Brazzaville"
          },
          "defaultGraceDays": {
            "type": "number",
            "minimum": 0,
            "maximum": 60
          },
          "receiptFooterText": {
            "type": "string",
            "nullable": true
          },
          "whatsappEnabled": {
            "type": "boolean"
          },
          "smsEnabled": {
            "type": "boolean"
          }
        }
      },
      "MemberUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "phone": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          }
        },
        "required": ["id", "phone", "fullName"]
      },
      "MemberDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "user": {
            "$ref": "#/components/schemas/MemberUserDto"
          },
          "role": {
            "type": "string",
            "enum": ["OWNER", "MANAGER", "ACCOUNTANT", "COLLECTOR", "VIEWER"]
          },
          "status": {
            "type": "string",
            "enum": ["ACTIVE", "SUSPENDED"]
          },
          "joinedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": ["id", "user", "role", "status", "joinedAt"]
      },
      "MemberListDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MemberDto"
            }
          }
        },
        "required": ["items"]
      },
      "UpdateMemberRoleDto": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "enum": ["OWNER", "MANAGER", "ACCOUNTANT", "COLLECTOR", "VIEWER"]
          }
        },
        "required": ["role"]
      },
      "InvitationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "phone": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": ["OWNER", "MANAGER", "ACCOUNTANT", "COLLECTOR", "VIEWER"]
          },
          "status": {
            "type": "string",
            "enum": ["PENDING", "ACCEPTED", "EXPIRED", "REVOKED"]
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": ["id", "phone", "role", "status", "expiresAt", "createdAt"]
      },
      "InvitationListDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvitationDto"
            }
          }
        },
        "required": ["items"]
      },
      "CreateInvitationDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "example": "+242066000002"
          },
          "role": {
            "type": "string",
            "enum": ["OWNER", "MANAGER", "ACCOUNTANT", "COLLECTOR", "VIEWER"]
          },
          "fullName": {
            "type": "string",
            "example": "Alphonse Ngoma"
          }
        },
        "required": ["phone", "role"]
      },
...
    }
  }
}